+++ CHANGELOG.md
@@ -4,7 +4,7 @@
 
 This release includes the following breaking changes
  * 🔥 The `strings.json` file format now maps keys to `{ value: "", comment: "" }` to support string comments (see below for details)
- * 🔥 The `sounds.json` file has renamed the `eax` field to `effect`
+ * 🔥 The `sounds.json` file has renamed the `eax` field to `effect`, and the `flags` adds a new `imported` boolean
 
 The compiled target version is now `es2015`, up from `es5` (Node has supported the vast majority of `es2015` features since around v6). This allows the codebase to take advantage of new TypeScript features.